Move product-detail to app directory (#152)
* Move product-detail to app directory

* change product-detail folder name on pages directory

* Remove unnecessary fragment

* renamed products url

* remove legacy product page

* refactor product not found functionality

* cleanup

* build fix

---------
