+++ src/components/product-detail/ProductImageGallery.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import ImageGallery from 'react-image-gallery';
 
 import 'react-image-gallery/styles/css/image-gallery.css';
@@ -11,6 +13,7 @@ interface ProductProp {
 const ProductImageGallery = ({ product }: ProductProp) => {
   const viewportWidth = useViewportWidth();
   const isMobile = viewportWidth <= 768;
+
   const images: {
     original: string;
     thumbnail: string;
@@ -28,22 +31,20 @@ const ProductImageGallery = ({ product }: ProductProp) => {
   });
 
   return (
-    <>
-      <div id="product-gallery" className="h-full w-full">
-        <ImageGallery
-          showThumbnails={isMobile ? false : true}
-          items={images}
-          thumbnailPosition={isMobile ? 'bottom' : 'left'}
-          showPlayButton={false}
-          showNav={false}
-          showBullets={isMobile ? true : false}
-          isRTL={false}
-          useTranslate3D={false}
-          useBrowserFullscreen={true}
-          showFullscreenButton={true}
-        />
-      </div>
-    </>
+    <div id="product-gallery" className="h-full w-full">
+      <ImageGallery
+        showThumbnails={isMobile ? false : true}
+        items={images}
+        thumbnailPosition={isMobile ? 'bottom' : 'left'}
+        showPlayButton={false}
+        showNav={false}
+        showBullets={isMobile ? true : false}
+        isRTL={false}
+        useTranslate3D={false}
+        useBrowserFullscreen={true}
+        showFullscreenButton={true}
+      />
+    </div>
   );
 };
