+++ src/components/product-detail/ProductInfo/ProductTitle.tsx
@@ -3,8 +3,8 @@ type TitleProps = {
   id?: string;
 };
 
-const ProductTitle = ({ title }: TitleProps) => {
-  return <h5 className="text-xl font-quicksand uppercase">{title}</h5>;
-};
+const ProductTitle = ({ title }: TitleProps) => (
+  <h5 className="text-xl font-quicksand uppercase">{title}</h5>
+);
 
 export default ProductTitle;
