+++ src/components/product-detail/RelatedProducts.tsx
@@ -1,4 +1,5 @@
 import ProductList from '../ProductList';
+
 interface ProductProp {
   product: Product;
   products: Product[];
@@ -7,14 +8,19 @@ interface ProductProp {
 
 const RelatedProducts = ({ title, product, products }: ProductProp) => {
   const relatedProducts = products.filter((prod) => prod.id !== product.id); // Remove the current product from the list of related products
-  return relatedProducts.length > 0 ? (
+
+  if (relatedProducts.length === 0) {
+    return null;
+  }
+
+  return (
     <>
       <div className="flex flex-col text-center mt-12 mb-16">
         <p className="text-3xl py-4 font-libre">{title}</p>
       </div>
       <ProductList relatedProducts={true} products={relatedProducts} />
     </>
-  ) : null;
+  );
 };
 
 export default RelatedProducts;
